Move dataflow template provider as subclass of parent dataflow provider
