+++ perfkitbenchmarker/configs/benchmark_config_spec.py
@@ -141,6 +141,7 @@ class _DpbServiceSpec(spec.BaseSpec):
                     dpb_service.DATAPROC_GKE,
                     dpb_service.DATAPROC_SERVERLESS,
                     dpb_service.DATAFLOW,
+                    dpb_service.DATAFLOW_TEMPLATE,
                     dpb_service.EMR,
                     dpb_service.GLUE,
                     dpb_service.UNMANAGED_DPB_SVC_YARN_CLUSTER,
