+++ perfkitbenchmarker/dpb_service.py
@@ -75,6 +75,7 @@ DATAPROC = 'dataproc'
 DATAPROC_GKE = 'dataproc_gke'
 DATAPROC_SERVERLESS = 'dataproc_serverless'
 DATAFLOW = 'dataflow'
+DATAFLOW_TEMPLATE = 'dataflow_template'
 EMR = 'emr'
 GLUE = 'glue'
 UNMANAGED_DPB_SVC_YARN_CLUSTER = 'unmanaged_dpb_svc_yarn_cluster'
